+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.2.0
+
+- Added option to provide pattern to search for:
+
+  ```
+  hotserve [mainHtmlFile] [port] [pattern]
+  ```
+- Added a startup check to error if watchman hasn't been installed
+- Added option to fetch a list of folders and files using /files endpoint
+
 ## 1.1.3
 
 ### 2020-06-12